refactor: lift `types` modules into crate root
