+++ securedrop-protocol/protocol-minimal/src/ciphertext.rs
@@ -1,4 +1,4 @@
-use super::constants::*;
+use crate::constants::*;
 use alloc::vec::Vec;
 use anyhow::Error;
