--- securedrop-protocol/protocol-minimal/src/types/mod.rs
@@ -1,23 +1,0 @@
-mod ciphertext;
-mod constants;
-mod journalist;
-mod key_types;
-mod source;
-mod traits;
-
-pub use constants::{LEN_DH_ITEM, LEN_MLKEM_ENCAPS_KEY, LEN_XWING_ENCAPS_KEY};
-
-pub use ciphertext::{CombinedCiphertext, Envelope, FetchResponse, Plaintext};
-
-pub use key_types::{
-    DhAkemKeyPair, DhFetchKeyPair, Enrollment, KeyBundlePublic, KeyPair, MlKem768KeyPair,
-    SessionStorage, SignedKeyBundlePublic, SignedLongtermPubKeyBytes, SigningKeyPair, XWingKeyPair,
-};
-
-pub use traits::{Enrollable, JournalistPublic, UserPublic, UserSecret};
-
-pub use journalist::{Journalist, JournalistPublicView};
-pub use source::{Source, SourcePublicView};
-
-pub(crate) use key_types::MessageKeyBundle;
-pub(crate) use traits::private;
