+++ securedrop-protocol/protocol-minimal/tests/core.rs
@@ -10,9 +10,7 @@ use securedrop_protocol_minimal::keys::{FPFKeyPair, NewsroomKeyPair};
 use securedrop_protocol_minimal::messages::core::SourceJournalistKeyRequest;
 use securedrop_protocol_minimal::primitives::MESSAGE_ID_FETCH_SIZE;
 use securedrop_protocol_minimal::server::Server;
-use securedrop_protocol_minimal::types::{
-    Journalist, JournalistPublic, Source, UserPublic, UserSecret,
-};
+use securedrop_protocol_minimal::{Journalist, JournalistPublic, Source, UserPublic, UserSecret};
 
 // TODO: better way (eg parameterize as in benchmarks)
 pub const DEFAULT_NUM_EPHEMERAL_KEYBUNDLES_JOURNALIST: usize = 3;
